requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-middleware==1.4.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
pipenv==2018.11.26
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
